backend/requirements.txt
fastapi==0.120.4
uvicorn[standard]==0.38.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
alembic==1.17.1
python-dotenv==1.2.1
pydantic[email]==2.12.3
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
email-validator==2.1.0